@@ -1,8 +1,8 @@
 package com.forum.domain;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 
 /*@Entity
 @Table(name="users", indexes = {
