@@ -7,7 +7,7 @@ import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import javax.validation.Constraint;
+import jakarta.validation.Constraint;
 
 @Constraint(validatedBy=RetypePasswordValidator.class)
 @Documented
