@@ -3,15 +3,15 @@ package com.forum.commands;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
-
 import org.springframework.web.multipart.MultipartFile;
 
 import com.forum.domain.Opinion;
 import com.forum.domain.Theme;
 import com.forum.domain.User;
+
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 
 
 public class OpinionCommand {
