@@ -1,16 +1,11 @@
 package com.forum;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+// @RunWith(SpringRunner.class)
+// @SpringBootTest
+// public class ForumApplicationTests {
 
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ForumApplicationTests {
+// 	// @Test
+// 	// public void contextLoads() {
+// 	// }
 
-	@Test
-	public void contextLoads() {
-	}
-
-}
+// }
